telegram/requirements.txt
# Telegram 模块依赖

# HTTP 请求
requests>=2.32.0

# WebSocket 客户端
websockets>=13.0

# 环境变量管理
python-dotenv>=1.0.0

